+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed: ValueError: too many values to unpack (expected 2)
 - Cleanup will now delete Jellyseerr accounts as well
+- Added Trial jellyfin account support (enable in .env). Will not create a Jellyseerr account, lasts 24 hours, one time use.
 
 # 1.0.2